Specifies a default playlist save path
Specifies a default playlist save path, it will be Your-MPV-configuration-direction/playlists. And checks if the custom save-path is available, and if not, uses the default path. If the default path does not exist, it will be created automatically (I'm not sure if it's a reasonable behavior to create folders on someone's computers without their confirmation)